ci: Fix the commit and repo spell checker
Rework the commit linter job to actually spell check commit
messages as well. The issue is that it needs a plugin
that the action is not taking into account, so we revert to
running the actual linter locally. To do this, we need to
install the plugin in the runner.

Also, we add an extra option to separate the footer from the
body, as the default parsing for `commitlint` assumes only the
`BREAKING CHANGE` as the separation keyword of the body and footer.

Additionally, we skip lines containing backticks, or triple backticks
(for code fences). This is a minor downside, but it is nice to be
able to include example code blocks in commit messages to showcase
the functionality of the patch. If we want to include terms that are
not in common dictionaries in commit messages, we should include them
in backticks.

Finally, we remove typos and replace it with `cspell`, as it seems
a bit better in catching typos.

PR: https://github.com/urunc-dev/urunc/pull/228
Signed-off-by: Anastassios Nanos <ananos@nubificus.co.uk>
Reviewed-by: Georgios Ntoutsos <gntouts@nubificus.co.uk>
Approved-by: Georgios Ntoutsos <gntouts@nubificus.co.uk>