refactor: Export BytesToStringMB
The BytesToStringMB function will be useful for the intergation of
virtiofs in urunc. In particular, we cna use it to convert the memory
limit of the container to MB and then string in order to xlaculate the
size of the tmpfs mount where the shared file between qemu and virtiofs
will reside.

PR: https://github.com/urunc-dev/urunc/pull/288
Signed-off-by: Charalampos Mainas <cmainas@nubificus.co.uk>
Reviewed-by: Anastassios Nanos <ananos@nubificus.co.uk>
Reviewed-by: Georgios Ntoutsos <gntouts@nubificus.co.uk>
Approved-by: Anastassios Nanos <ananos@nubificus.co.uk>