fix(linux): Use a file to pass environment variables in urunit
There are cases where containers use non-POSIX environment variables or
cases where the environment variables can exceed the limit of characters
in Linux kernel boot parameters. Therefore, the environment variables
were not set correctly.

To fix the above issue, we change the way we pass the environment
variables to the guest in Linux with the the help of urunit. In
particular, instead of using the Linux kernel boot parameters, we use
store the environment variables in a file and let urunit read that file
and set the environment variables for the application inside the Linux
VM.

In order to minimize the depedencies of the guest kernel and provide a
croos-monitor solution, we pass the urunit configuration file as an
initrd for the guest and instruct Linux kernel to keep it with the
retain_initrd kernel boot parameter. In that way, urunit simply reads
the contents of /sys/firmware/initrd and is able to get the
configuration file.

If the guest is configured to use an initrd then we format the urunit
configuration file in the cpio format and concatenate the user provided
initrd with the one we created for the urunit configuration file. The
Linux kernel is able to properly extract concatenated initrds and
therefore the configuration file will appear in the rootfs of the VM.

PR: https://github.com/urunc-dev/urunc/pull/308
Signed-off-by: Charalampos Mainas <cmainas@nubificus.co.uk>
Reviewed-by: Anastassios Nanos <ananos@nubificus.co.uk>
Approved-by: Anastassios Nanos <ananos@nubificus.co.uk>