fix(firecracker): Allow guests define args for firecracker
The guest can append further cli args for the monitor, since the VMM
interface calls the MonitorCli function. However, this function was not
compatible with Firecracker, which uses a json configuration file to
define the VM arguments.

THis commit partially fixes this by updating the return value of
MonitorCli, in order to use a new struct. The current implementation was
also necessary for Linux to define the extra initrd file with the urunit
configuration. Therefore, for the time being the MonitrCliArgs consists
of only two fields, one for initr and one for other cli args. However,
in the future we can iterate over it and update it.

PR: https://github.com/urunc-dev/urunc/pull/308
Signed-off-by: Charalampos Mainas <cmainas@nubificus.co.uk>
Reviewed-by: Anastassios Nanos <ananos@nubificus.co.uk>
Approved-by: Anastassios Nanos <ananos@nubificus.co.uk>