feat(block rootfs): Copy files from the mount list in the VM's rootfs
Since we cna not bind mount files into a mounted block image, we copy
them instead and therefore update the block image with the new files.
Hence, when we attach the block image to the VM, it will contain the
files from the mount configuration of the container. However, we still
need to find a solution for directories.

PR: https://github.com/urunc-dev/urunc/pull/269
Signed-off-by: Charalampos Mainas <cmainas@nubificus.co.uk>
Signed-off-by: Anastassios Nanos <ananos@nubificus.co.uk>
Reviewed-by: Anastassios Nanos <ananos@nubificus.co.uk>
Approved-by: Anastassios Nanos <ananos@nubificus.co.uk>