fix(copyFile): Define the full destination path in copyFile arguments
Make copyFile more generic in order to allow copying files using a
different filename than the ne in the sourc path. For example, copying
file /a/b/c/file1 to /e/f/g/file2 was not possible, since the old
implementation was using the file1 as the basename for the new file.

PR: https://github.com/urunc-dev/urunc/pull/269
Signed-off-by: Charalampos Mainas <cmainas@nubificus.co.uk>
Signed-off-by: Anastassios Nanos <ananos@nubificus.co.uk>
Reviewed-by: Anastassios Nanos <ananos@nubificus.co.uk>
Approved-by: Anastassios Nanos <ananos@nubificus.co.uk>