feat(linux): set/unset urunit environment variable for default route
In the case of Linux guests, urunit by default used to set as the
default route the eth0 interface. However, this was creating issues with
docker and nerdctl, where the container could not reach the outside
network. Nevertheless, this was required in k8s set ups, since the IP of
the container could be in a different subnet than the gateway and the
Linux kernel can not handle such network configuration.

To fix the above, use a specific environment variable for urunit, called
`URUNIT_DEFROUTE`, which urunc sets it only when the IP of the container
is in a different subnet than the gateway. Otherwise urunc will not set
it. From urunit's side, if the environment variable is set, then it will
use the eth0 interface as the default route.

PR: https://github.com/urunc-dev/urunc/pull/284
Signed-off-by: Charalampos Mainas <cmainas@nubificus.co.uk>
Reviewed-by: Anastassios Nanos <ananos@nubificus.co.uk>
Approved-by: Anastassios Nanos <ananos@nubificus.co.uk>