refactors(vmm): move interface and structs to types
SImilarly to the unikernel package, move the vmm interface and the data
types used for its initialization to the types package. Therefore, we
will not have problem with depedency cycles and we can reference these
types any where in the unikontainers, unikernels or vmm package.

PR: https://github.com/urunc-dev/urunc/pull/306
Signed-off-by: Charalampos Mainas <cmainas@nubificus.co.uk>
Reviewed-by: Anastassios Nanos <ananos@nubificus.co.uk>
Approved-by: Anastassios Nanos <ananos@nubificus.co.uk>