refactor(block): unify the handling of block images from any source
Create a new function to handle the case where a block image is present
inside the container rootfs. Furthermore, unify the handling of such
cases with the case where we check and use the container rootfs as a
block image for the guest's rootfs.

PR: https://github.com/urunc-dev/urunc/pull/306
Signed-off-by: Charalampos Mainas <cmainas@nubificus.co.uk>
Reviewed-by: Anastassios Nanos <ananos@nubificus.co.uk>
Approved-by: Anastassios Nanos <ananos@nubificus.co.uk>